# Make location a struct