--- a/tests/golden/benchmark_test.go
+++ b/tests/golden/benchmark_test.go
@@ -47,13 +47,13 @@ func BenchmarkGolden(b *testing.B) {
 			if err := json.Unmarshal(data, &input); err != nil {
 				b.Fatal(err)
 			}
+			model, err := factory.NewModel(input, factory.Options{})
+			if err != nil {
+				b.Fatal(err)
+			}
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
 				b.StopTimer()
-				model, err := factory.NewModel(input, factory.Options{})
-				if err != nil {
-					b.Fatal(err)
-				}
 				solver, err := nextroute.NewParallelSolver(model)
 				if err != nil {
 					b.Fatal(err)