@dirkschumacher 2024-05-14
> I would not do this. This will allow users to update a Location without us having any control on the values set. Today one can only create a location by invoking the factory method NewLocation which makes it very clear where values are set.

That's not true as the fields are private

```go
type Location struct {
	longitude float64
	latitude  float64
}
```