@dirkschumacher 2024-05-14
> Would be nice if this PR is actually also leveraging this change in the nextroute code.

@davidrijsman can you elaborate a bit what you mean?