@dirkschumacher 2024-05-17
For now I leave the size of the struct as is. It appears that this is the faster approach during the search as most time is spent on validating locations. We can revisit that later.